builtins to one call, updated range
all the builtins are now executed at the beginning under one initialize call. All further builtins should be added to that function to be added to the builtin dictionary on startup. range now has an implemented function so that the translation now works correctly. updated the new syntax to fit this change.