whileAst printSmalltalk created, tryAst added more generality, raiseAst added from functionality
Created the printSmalltalk function in whileAst so it can properly translate python while loops to smalltalk. Added error catching in addition to exception checking to better mirror the general tryException functionality in python. Added the ability to translate raise x from y to small talk and added tests to check if it is working. (may need fixed later but currently functioning).