added more variables helpers and testing it
redesigned the read method to read from all previous layers. Also redesigned both methods so it could function properly as a global variable. Also implemented two tests to make sure that both methods and add and read from AllVariables properly. Also edited the functionDef for later updating while it was fresh in my head.  Also changed the inheritance of the helper object to an object.