moved vars to part of Variables class
I removed all of the references to vars in all of the programs and made a helperSymbols instance variable in Variables. This allowed me to remove a lot of inlined, confusing, deletes to vars and methods that required including vars. This reduced the messages needed and the arguments passed by a lot.