added all zeroDivide errors, updated list __repr__,  fixed tests
added dividing by zero errors to the float, int, and complex numbers. Refactored the __repr__ method in list to actually return a string instead of error or return a str(). required an update in the str __repr__ as well. fixed tests accordingly. Fixed slice so that it checks for a NoneType rather than the word None. Changed a bunch of tests to the python error type and made their functions throw that error.