repr and str functions and __repr__ for range
I forgot to implement the repr and str actual builtin functions when I was working on changing all of those. I also realized that range needed a different implementation for __repr__ since it only tells the information the function was called with. This meant I needed two new variables in the range class so I could preserve the call values. I also created tests for all three of these features.