changed all = to == where applicable, added dictionary tests, change args to params and opp, fixed defaults,  int handles strings
update = to == for a faster execution time where singletons or class names are being used. Added tests to the data structure conversion functions to check for dictionary conversion. Changed args to params and params to args in the FunctionDef class so that they were appropriate with the actual definition. The int function is now capable of converting binary, octal, decimal, and hexidecimal strings to integers.