sum function, add fix, more errors
created a the sum() builtin python function. This required fixing the add functionality for integers, floats, and complex numbers so that types are respected properly. Also I saw that there wasn't correct error handling for a bunch of the types. I fixed the add and multiply functionality in int, float, and complex but I am adding the rest to my stretch goals.