locals(), globals(), round(), start of pow()
implemented the local and global builtin functions and a method in variables to return the a dictionary of the variables for them. Implemented the round builtin function and a python object method for the error checking. Started work on pow(), this is requiring a __pow__ method in all numbers and one in objects for error checking. This is using the double dispatch design pattern. Also in the process of changing __add__ over to double dispatch as well.