pow() function completed
I updated all of the  ___powReal:imag: methods. I changed the int version so that it was more efficient. I updated the float version so it was the previous version of the in method. I updated all of the ones for complex values too. The ones for int and float now do the basic Euler's method conversion.  I also updated the complex version so that it turned the operand into euler's form, multiplied the exponent of the new form and the actual exponent, and then used the float method to break the resulting method back into a solution to the exponential. I'm a freaking genius. Hallelujah I'm done with this function. Complex numbers can eat me.