dict() revert, pow rename, test revamp, double dispatch add and mul
Somehow dict got reverted to a previous version so I restored it to working order, the pow parser uses __pow__ so I changed all ___pow methods to __pow__: . I added tests to the __pow__ methods for each class so more cases are checked. Finished converting the add method to double dispatch and completed the __mul__ method to be double dispatch.