continued work on string format, update container equality,  added tuple parser
Continued work on adding the string formatting for modulo.  Starting with the string version and trying to create it as double dispatching. Made an implementation to identify single % but still replace %% with a single % in the final output string. Also since this method depends heavily on tuples I create the ability to parse tuples so I can test it better later. Also because I was looking at tuples I put in a line in the equality statement to make sure lists can't equal tuples.