implemented Gaussian rounding, string injection now works again, added possible other working injection
Python uses Gaussian rounding and I didn't know it. Updated the round function and class methods to use Gaussian rounding. I also updated the Format tag so now it is completely functional. Now I am implementing the class formatters to handle the flags, precision, and type checking. The string one is fully implemented and manually tests. The list, set, frozenset, dictionary, range, and tuple also have an implementation because it basically the same as strings but I haven't tested it yet. Also there is a standard version for all objects which probably does not work properly.