full logic implementation and fix for indentation printing
I have fully implemented the logical operators and, or, and not so that they both exist and work on non-boolean types, using their associated __bool__ method. Also implemented the parser for and and or. This also involved a not particularly elegant, but effective, fix for the smalltalkSourceFor:... method not printing indentation correctly. I also found something wrong with the white space printing in the for loop printSmalltalkOn: method so I fixed that. I also made tests for and and or and updated the not tests for new data types.