basic parse change
Fixed a but where in a for loop the iterator was not turned into a python variable for accessing. Also removed the end . in the for loop parser because for loops are statements and will automatically have one added. Also fixed a bug in a test where a 'hidden' tab was not present.